Add beginning of aiu: AppImageUtilities
I figure the AppImage standard is fairly simple, so I might as well make
some utilities for it for updating, listing, etc.

Currently, the only two commands that are implemented are `aiu list` and
`aiu check`.

And of course I have a script to package these utilities into an
AppImage.